requirements.txt
Flask>=2.3.7
pymongo==4.5.0
python-dotenv==1.0.0
werkzeug>=2.3.7
flask-cors==3.0.1
